Remove Docker and install PHP on host
Docker is increasingly causing issues, particularly around networking
and maintenance.  With only a couple of sites now hosted on this
machine, it will be much easier to simply install PHP and Composer, and
manually setup and maintain these sites via the command-line.

Install PHP 8.0 with the required extensions and Composer.  This should
be all that is needed, with Nginx already being installed and configured
and MariaDB being hosted on another machine.